requirements.txt
fastapi==0.115.14
python-dotenv==1.1.1
sqlalchemy==2.0.41
colored==2.3.0
psycopg2==2.9.10
pydantic==2.11.7
boto3==1.39.1

#for debug. remove for release
uvicorn==0.35.0